requirements.txt
pandas~=2.1.0
selenium~=4.12.0
fitz~=0.0.1.dev2
pytesseract~=0.3.10
Pillow~=10.0.0
pathlib~=1.0.1
requests~=2.31.0
xlrd~=2.0.1
pdfquery~=0.4.3
beautifulsoup4~=4.12.2
pypdf2~=3.0.1

numpy~=1.25.2
tika~=2.6.0
langdetect~=1.0.9